requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
boto3==1.42.11
    # via -r requirements.in
botocore==1.42.28
    # via
    #   boto3
    #   s3transfer
certifi==2026.1.4
    # via
    #   elastic-transport
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via
    #   -r requirements.in
    #   sendgrid
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-sendgrid-v5
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers>=4.0
    # via -r requirements.in
django-elasticsearch-dsl==8.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-sendgrid-v5==1.3.0
    # via -r requirements.in
django-storages==1.14.6
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-yasg==1.21.7
    # via -r requirements.in
elastic-transport==8.17.1
    # via
    #   elasticsearch
    #   elasticsearch-dsl
elasticsearch==8.16.0
    # via elasticsearch-dsl
elasticsearch-dsl==8.16.0
    # via django-elasticsearch-dsl
forex-python==1.8
    # via python-paystack
gunicorn==23.0.0
    # via -r requirements.in
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpickle==4.1.1
    # via python-paystack
markupsafe==3.0.3
    # via werkzeug
packaging==25.0
    # via
    #   drf-yasg
    #   gunicorn
pillow==11.1.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.23
    # via cffi
pycryptodome==3.22.0
    # via -r requirements.in
pydantic==2.12.5
    # via -r requirements.in
pydantic-core==2.41.5
    # via pydantic
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   elasticsearch-dsl
python-http-client==3.3.7
    # via
    #   django-sendgrid-v5
    #   sendgrid
python-paystack==1.2.0
    # via -r requirements.in
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.3
    # via drf-yasg
requests==2.32.3
    # via
    #   -r requirements.in
    #   forex-python
    #   python-paystack
s3transfer==0.16.0
    # via boto3
sendgrid==6.12.5
    # via django-sendgrid-v5
simplejson==3.20.2
    # via forex-python
six==1.17.0
    # via
    #   django-elasticsearch-dsl
    #   python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.14.1
    # via
    #   dj-database-url
    #   elasticsearch-dsl
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via django
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.6.3
    # via
    #   botocore
    #   elastic-transport
    #   requests
validators==0.35.0
    # via python-paystack
werkzeug==3.1.5
    # via sendgrid
whitenoise==6.11.0
    # via -r requirements.in
